batch/pyproject.toml
[project]
name = "batch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-storage>=3.10.1",
    "jupyter>=1.1.1",
    "pandas>=3.0.1",
    "pyspark>=4.1.1",
    "requests>=2.32.5",
]
